Cast to float

+++ policyengine_uk_data/utils/loss.py
@@ -367,13 +367,15 @@ def create_target_matrix(
     benefit_cap_reduction = sim.calculate(
         "benefit_cap_reduction", map_to="household"
     ).values
-    df["dwp/benefit_capped_households"] = benefit_cap_reduction > 0
+    df["dwp/benefit_capped_households"] = (benefit_cap_reduction > 0).astype(
+        float
+    )
     target_names.append("dwp/benefit_capped_households")
     target_values.append(
         115_000
     )  # https://www.gov.uk/government/statistics/benefit-cap-number-of-households-capped-to-february-2025/benefit-cap-number-of-households-capped-to-february-2025
 
-    df["dwp/benefit_cap_total_reduction"] = benefit_cap_reduction
+    df["dwp/benefit_cap_total_reduction"] = benefit_cap_reduction.astype(float)
     target_names.append("dwp/benefit_cap_total_reduction")
     target_values.append(
         60 * 52 * 115_000
